overlay/requirements.txt
pillow>=10.0.0
websockets>=12.0
